Update compiled protobuf to be Python 3 compatible.
This was done by just re-running a later version of the compiler over the proto
file again, but only taking the changes related to Python 3 compatibility.
There were some changes to the protobuf library itself, but python3_protobuf is
only up to version 2.5, whereas the compiler was 2.6. (And the two are not
directly compatible.) This is a significantly hacky attempt to get a compiled
protobuf that works in Python 2 & 3.